# openapi.yaml
openapi: 3.0.3
info:
  title: SalesMind API
  description: Mental fitness platform for sales teams - API for managing user profiles, assessments, skills, drills, and progress tracking
  version: 1.0.0
  contact:
    name: SalesMind Support
    email: support@salesmind.com
servers:
  - url: https://api.salesmind.com/v1
    description: Production server
  - url: https://staging-api.salesmind.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /drills:
    get:
      tags:
        - Drills
      summary: List all drills
      description: Retrieve a list of all available training drills
      security:
        - BearerAuth: []
      parameters:
        - name: skill_id
          in: query
          description: Filter drills by skill
          schema:
            type: string
        - name: type
          in: query
          description: Filter by drill type
          schema:
            type: string
            enum: [form, timer, interactive, tracking]
        - name: difficulty_level
          in: query
          description: Filter by difficulty
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
      responses:
        '200':
          description: Drills retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Drill'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /drills/{drillId}:
    get:
      tags:
        - Drills
      summary: Get drill details
      description: Retrieve detailed information about a specific drill
      security:
        - BearerAuth: []
      parameters:
        - name: drillId
          in: path
          required: true
          schema:
            type: string
            example: abc-card
      responses:
        '200':
          description: Drill details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drill'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /drill-completions:
    get:
      tags:
        - Drills
      summary: Get user's drill completions
      description: Retrieve the authenticated user's drill completion history
      security:
        - BearerAuth: []
      parameters:
        - name: skill_id
          in: query
          description: Filter by skill
          schema:
            type: string
        - name: start_date
          in: query
          description: Filter completions from this date
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: Filter completions until this date
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Drill completions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DrillCompletion'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Drills
      summary: Record drill completion
      description: Submit a completed drill session
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - drill_id
                - skill_id
                - duration_seconds
              properties:
                drill_id:
                  type: string
                  example: abc-card
                skill_id:
                  type: string
                  example: emotional-resilience
                duration_seconds:
                  type: integer
                  minimum: 0
                  example: 480
                score:
                  type: integer
                  minimum: 0
                  maximum: 100
                  example: 85
                completion_data:
                  type: object
                  description: Additional data specific to the drill
                  example:
                    activating_event: "Client rejected my proposal"
                    belief: "I'm not good enough"
                    consequence: "Felt discouraged"
      responses:
        '201':
          description: Drill completion recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrillCompletion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /progress/weekly:
    get:
      tags:
        - Progress
      summary: Get weekly progress
      description: Retrieve weekly progress statistics
      security:
        - BearerAuth: []
      parameters:
        - name: week_offset
          in: query
          description: Number of weeks to offset from current week (0 = current week, 1 = last week, etc.)
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Weekly progress retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  week_start:
                    type: string
                    format: date
                  week_end:
                    type: string
                    format: date
                  daily_progress:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        drills_completed:
                          type: integer
                        total_time_minutes:
                          type: integer
                        skills_practiced:
                          type: array
                          items:
                            type: string
                  total_drills_completed:
                    type: integer
                  total_time_minutes:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /progress/skill-improvements:
    get:
      tags:
        - Progress
      summary: Get skill improvements
      description: Retrieve skill score improvements over time
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          description: Time period for improvements
          schema:
            type: string
            enum: [week, month, quarter, year, all]
            default: month
      responses:
        '200':
          description: Skill improvements retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        skill_id:
                          type: string
                        skill_name:
                          type: string
                        initial_score:
                          type: integer
                        current_score:
                          type: integer
                        improvement:
                          type: integer
                        improvement_percentage:
                          type: number
                          format: float
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or signup
  schemas:
...
    Drill:
      type: object
      properties:
        id:
          type: string
          example: abc-card
        skill_id:
          type: string
          example: emotional-resilience
        title:
          type: string
          example: ABC Card Analysis
        description:
          type: string
          example: Identify and challenge negative thought patterns after rejection
        type:
          type: string
          enum: [form, timer, interactive, tracking]
          example: form
        component:
          type: string
          example: AbcCardDrill
        duration_minutes:
          type: integer
          example: 10
        difficulty_level:
          type: string
          enum: [beginner, intermediate, advanced]
          example: beginner
        is_active:
          type: boolean
          example: true
        completion_count:
          type: integer
          description: Number of times user has completed this drill
          example: 3
    DrillCompletion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        drill_id:
          type: string
          example: abc-card
        drill_title:
          type: string
          example: ABC Card Analysis
        skill_id:
          type: string
          example: emotional-resilience
        skill_name:
          type: string
          example: Emotional Resilience
        completed_at:
          type: string
          format: date-time
        duration_seconds:
          type: integer
          example: 480
        score:
          type: integer
          minimum: 0
          maximum: 100
          example: 85
        completion_data:
          type: object
          description: Drill-specific completion data
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        message:
          type: string
          example: The provided data is invalid
        status_code:
          type: integer
          example: 400
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Authentication token is missing or invalid
            status_code: 401
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: The requested resource does not exist
            status_code: 404